--- .gitignore
@@ -1,23 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-frontend/node_modules
-frontend/.pnp
-frontend/.pnp.js
-
-# testing
-/coverage
-
-# production
-frontend/build
-
-# misc
-frontend/.DS_Store
-frontend/.env.local
-frontend/.env.development.local
-frontend/.env.test.local
-frontend/.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
