--- frontend/src/App.js
@@ -1,7 +1,0 @@
-import "./App.css";
-
-function App() {
-  return <div className="App"> Tähän tulee climate äppi</div>;
-}
-
-export default App;
