+++ public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <div id="menu">
+    <div id="menu-items">
+      <a href='/' class="menu-item">User's Page</a>
+      <a href='/' class="menu-item">Login</a>
+      <a href='/' class="menu-item">Register</a>
+      <a href='/' class="menu-item">About</a>
+    </div>
+    <div id="menu-background-pattern"></div>
+    <div id="menu-background-image"></div>
+  </div>
+</html>
