+++ src/App.js
@@ -1,0 +1,16 @@
+import './App.css';
+
+const menu = document.getElementById("menu");
+
+Array.from(document.getElementsByClassName("menu-item"))
+  .forEach((item, index) => {
+    item.onmouseover = () => {
+      menu.dataset.activeIndex = index;
+    }
+});
+
+function App(){
+
+}
+
+export default App;
